requirements.txt
anthropic>=0.8.0
openai>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
markdown>=3.5.0
